rename agent collector into profile collector
