+++ docs/docs/api_ref/research_town.md
@@ -750,10 +750,10 @@
       * [`ReviewQualityEvaluator.eval()`](research_town.evaluators.md#research_town.evaluators.ReviewQualityEvaluator.eval)
 * [research_town.utils package](research_town.utils.md)
   * [Submodules](research_town.utils.md#submodules)
-  * [research_town.utils.agent_collector module](research_town.utils.md#module-research_town.utils.agent_collector)
-    * [`coauthor_filter()`](research_town.utils.md#research_town.utils.agent_collector.coauthor_filter)
-    * [`coauthor_frequency()`](research_town.utils.md#research_town.utils.agent_collector.coauthor_frequency)
-    * [`collect_publications_and_coauthors()`](research_town.utils.md#research_town.utils.agent_collector.collect_publications_and_coauthors)
+  * [research_town.utils.profile_collector module](research_town.utils.md#module-research_town.utils.profile_collector)
+    * [`coauthor_filter()`](research_town.utils.md#research_town.utils.profile_collector.coauthor_filter)
+    * [`coauthor_frequency()`](research_town.utils.md#research_town.utils.profile_collector.coauthor_frequency)
+    * [`collect_publications_and_coauthors()`](research_town.utils.md#research_town.utils.profile_collector.collect_publications_and_coauthors)
   * [research_town.utils.agent_prompter module](research_town.utils.md#module-research_town.utils.agent_prompter)
     * [`brainstorm_idea_prompting()`](research_town.utils.md#research_town.utils.agent_prompter.brainstorm_idea_prompting)
     * [`discuss_idea_prompting()`](research_town.utils.md#research_town.utils.agent_prompter.discuss_idea_prompting)
